Added build tags to build without unused features: - remote stores (boltdb and redis) - forwarding proxy
