+++ store_redis.go
@@ -1,3 +1,5 @@
+//+build !nostores
+
 /*
 Copyright 2015 All rights reserved.
 Licensed under the Apache License, Version 2.0 (the "License");
